INSANE

Insane Bayesian data augmentation

INSANE uses Bayesian data augmentation (DA) to perform inference on a number of evolutionary models on phylogenetic trees. As such, performing inference will output posterior samples for the governing parameters as well as complete or data augmented trees, that is, trees that include probable configurations of unobserved variables such as the lineages that went extinct in the past or the underlying (latent) speciation rate.

References

Birth-Death Diffusion (BDD): Quintero, I., Lartillot, N., Morlon, H. (2024). Imbalanced speciation pulses sustain the radiation of mammals. Science, 384: 1007-1012. link

Diffused Brownian motion (DBM): Quintero, I. (2025). The diffused evolutionary dynamics of morphological novelty. Proceedings of the National Academy of Sciences, U.S.A. 122 (18) e2425573122, link.

Insane tree and model input/output

Insane trees

INSANE uses different types of trees, but all are recursive structures. All trees have at least three fields:

d1 which corresponds to the left daughter (either another tree structure or nothing if it is a tip).

d2 which corresponds to the right daughter (also either another tree structure or nothing if it is a tip).

e a decimal number specifying the edge length of the tree.

For all tree output from simulations and inference (all expect labelled trees):

fx a Boolean stating if tree is fixed or not. That is, if it corresponds to a sampled (fx = true) or unsampled (fx = false) lineage.

For all trees that allow for extinction:

iμ a Boolean stating if tree is extinct (iμ = true) or not (iμ = false).

Insane simple trees (sT)

sT_label: is a simple labelled tree, and is the one required as input to perform inference. Note that all other model specific tree types can be converted to sT_label by using the latter as a function: tree = sT_label(tree_of_other_type).

It has the additional field l for label.

sTpb: is a simple pure-birth tree for the constant pure-birth (Yule) model.

sTbd: is a simple birth-death tree for the constant birth_death model.

Insane BDD trees (iT)

All iT trees specify to birth-death diffusion trees with different extinction assumptions and all hold the following extra fields:

dt a decimal specifying the time step of the GBM discretization.

fdt a decimal specifying the final time step of the GBM discretization.

lλ an array specifying the Brownian motion evolution of log-speciation rates.

And have the following concrete types:

iTpb, iTce and iTct: are BDD trees with no-extinction (pure-birth), constant extinction and constant turnover, respectively.

iTbd: is the full BDD tree.

It has the additional field lμ: an array specifying the Brownian motion evolution of log-extinction rates.

Insane DBM trees (sTxs)

dt a decimal specifying the time step of the GBM discretization.

fdt a decimal specifying the final time step of the GBM discretization.

xv an array specifying the Brownian motion evolution of traits.

lσ2 an array specifying the Geometric Brownian motion evolution of rates.

Reading and saving newick trees

Tapestree can read files in the simple newick format using the read_newick function:

tree = read_newick(joinpath(dirname(pathof(Tapestree)), "..", "data", "tree_50.tre"))

Note that the tree has type sT_label, which stands for simple labelled tree. You can check this using

typeof(tree)

Similarly, it can also write trees using write_newick

write_newick(tree, "<directory>")

Note that only *_label trees (e.g., sT_label) have labels. So if you want to save a DA tree with the original tip labels plus new names to the data augmented trees, you first will have to create a labelled tree from the DA_tree and the loaded labelled tree and then save it:

write_newick(sT_label(DA_tree, tree), "<directory>")

Reading and saving model output

All models can return and/or save the output by writing directly to a file on the fly and within the Julia session when the model finishes. There are two outputs:

The governing parameters trace, which is saved as a .log file and returned as the first object once the algorithm finished. These can be conveniently read in the Tracer software (https://github.com/beast-dev/tracer/releases/tag/v1.7.2)[https://github.com/beast-dev/tracer/releases/tag/v1.7.2].

The DA trees as a tree vector which are saved as a .txt file (of the same name as the .log) and returned as the second object once the algorithm finished.

The DA trees written in the insane-specific .txt file can be read using the iread() function, which only needs the specific file directory as input, but also accepts an optional (keyword) argument ix that indicates the specific tree iterations to read, as an OrdinalRange object. For instance, to read only the first $50$ trees, one can use iread("<directory to txt>", ix = 1:50. To read only the trees every $10$ iterations from the $100$ to $400$ sampled, one can use ix = 100:10:400, and so on. This can be helpful to avoid high computation costs of reading very large files.

You can save any individual or vector of insane trees using the iwrite() function: iwrite(trees, "<directory>"), which can be read by Tapestree. For portability, you can also save the DA trees as nexus files where nodes have been annotated with the vector of rates using write_nexus(trees, tree, "<directory>"), where trees is the tree vector and tree is the original labelled tree (i.e., the one you used as input into an insane model).

Full documentation

Tapestree.INSANE.read_newick — Function

read_newick(in_file::String; ix::OrdinalRange{Int64,Int64} = 0:0)

Reads a newick tree into sT from in_file at lines ix.

read_newick(in_file::String,
            fossil ::Bool;
            ix     ::OrdinalRange{Int64,Int64} = 0:0,
            ne     ::Float64                   = accer)

Reads a newick tree into sTf from in_file at lines ix.

Tapestree.INSANE.write_newick — Function

write_newick(tree::T, ofile::String)

Writes an iTree as a newick tree to ofile.

write_newick(treev::Vector{T}, ofile::String)

Writes an iTree as a newick tree to ofile.

Tapestree.INSANE.iread — Function

iread(file::String; ix::OrdinalRange{Int64,Int64} = 0:0)

Read a tree file exported by insane in file and with optional OrdinalRange specifying which trees to sample.

Tapestree.INSANE.iwrite — Function

iwrite(tree::T, ofile::String) where {T <: iTree}

Write iTree to file.

iwrite(tree::Vector{T}, ofile::String) where {T <: iTree}

Write a vector of trees.

Tapestree.INSANE.write_nexus — Function

write_nexus(tree::T, reftree::sT_label, ofile::String) where {T <: iTree}

Writes an iTree as a extensive nexus tree to ofile.

write_nexus(treev::Vector{T}, reftree::sT_label, ofile::String) where {T <: iTree}

Writes an iTree as a extensive nexus tree to ofile.

Insane models

Tree input

All inference functions require a phylogenetic tree of type sT_label, that is, a simple labelled tree. This is the default object type when using the read_newick function. However, when using simulations from models or iread, the resulting type is specific to the model (for computational efficiency). One can easily create a tree of type sT_label from any other tree by using sT_label(tree). The output can then be used to perform inference.

Common (keyword) arguments across all insane inference models

nburn: specifies the number of iterations to discard as burn-in.

niter: specifies the number of MCMC iterations.

nthin: specifies the iteration frequency at which to save the parameters in the julia session (i.e.,nthin = 2 specifies saving every 2 iterations),

nflush: specifies the frequency at which to save to file.

ofile: specifies the directory where the results will be written.

tρ: controls the sampling fraction and receives a Dictionary as input, with a String key pointing to a Float64 number (i.e., Dict{String, Float64}). If the dictionary is of length 1 with an empty string, then the insane sets this as a the global sampling fraction. For example, to set a sampling fraction of 0.6, one show input tρ = Dict("" => 0.6). Most times, however, sampling fraction is not uniform across the tree, but rather some part so the tree is more heavily sampled than others, to accommodate these variability, you can input a dictionary of the same length as the number of tips in the tree, where the dictionary key string is the tip label pointing to the specific sampling fraction value. For example, for two tips, named tip_1 and tip_2, one could input tρ = Dict("tip_1" => 0.5, "tip_2" => 0.3). Make sure to specify all tips when assigning different sampling fractions across the tips, even ones with 1.0.

prints: specifies the number of seconds to refresh the progress meter.

survival: For those modesl with extinction, true or false if to condition the likelihood on survival.

Constant rate models

Constant pure-birth (Yule) process (CPB)

The simplest diversification model assumes no extinction and a constant speciation rate $\lambda$, also known as, the pure-birth or Yule model.

Simulations

To simulate a pure-birth tree one can use sim_cpb. For instance, for a period of $10$ time units and a speciation rate of $\lambda = 0.5$:

tr = sim_cpb(10.0, 0.5)

Full documentation

Tapestree.INSANE.sim_cpb — Function

sim_cpb(t::Float64, λ::Float64)

Simulate a constant pure-birth iTree of height t with speciation rate λ.

sim_cpb(t::Float64, λ::Float64, na::Int64)

Simulate a constant pure-birth iTree of height t with speciation rate λ.

Inference

To perform inference on a tree (of type sT_label), we can use the insane_cpb function (cpb = constant pure-birth).

r, tv = insane_cpb(tree,
                   nburn  = 500,
                   niter  = 1_000,
                   nthin  = 2,
                   nflush = nthin,
                   ofile  = "<directory>")

Note that for this specific CPB model, where the sampling fraction, $\rho$, is $1$, there are no unobserved components since we assume no extinction and that all species have been sampled. In this case, all the trees in the tree vector will be exactly the same.

In the following example, we now specify a global sampling fraction of 0.8.

r, tv = insane_cpb(tree,
                   nburn  = 500,
                   niter  = 1_000,
                   nthin  = 2,
                   nflush = nthin,
                   ofile  = "<directory>",
                   tρ     = Dict("" => 0.8))

This time, the DA trees are different from one another since they have data augmented lineages that represent that proportion of species not included in the tree. The position change from tree to tree because we are integrating over their unknown placement. You can check this by plotting the trees (check Insane plots). For instance, to plot the first tree in the vector:

plot(tv[1])

Finally, it is important to note that we use Gibbs sampling across most parameter updates in INSANE. So, the prior for speciation in the CPB is a Gamma prior, and its parameters can be specified with argument λ_prior, The default is λ_prior = (1.0, 1.0).

Full documentation

Tapestree.INSANE.insane_cpb — Function

insane_cpb(tree    ::sT_label;
           λ_prior ::NTuple{2,Float64}     = (1.0, 1.0),
           niter   ::Int64                 = 1_000,
           nthin   ::Int64                 = 10,
           nburn   ::Int64                 = 200,
           nflush  ::Int64                 = nthin,
           ofile   ::String                = string(homedir(), "/cpb"),
           λi      ::Float64               = NaN,
           pupdp   ::NTuple{2,Float64}     = (0.2, 0.2),
           prints  ::Int64                 = 5,
           tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for constant pure-birth.

Constant birth-death process (CBD)

We now consider that species have an instantaneous constant rate of extinction, $\mu$, besides the constant rate of speciation $\lambda$.

Simulations

To simulate under the CBD model, say, for a period of $10$, with speciation rate of $0.5$, and extinction rate of $0.3$, a one can use

tr = sim_cbd(10.0, 0.5, 0.3)

Full documentation

Tapestree.INSANE.sim_cbd — Function

sim_cbd(t::Float64, λ::Float64, μ::Float64)

Simulate a constant birth-death iTree of height t with speciation rate λ and extinction rate μ.

sim_cbd(t ::Float64,
        λ ::Float64,
        μ ::Float64,
        na::Int64)

Simulate a constant birth-death iTree of height t with speciation rate λ and extinction rate μ.

Inference

To perform inference on the CBD model, we can use

r, tv = insane_cbd(tree,
                   nburn    = 1_000,
                   niter    = 50_000,
                   nthin    = 50, 
                   ofile    = "<directory>",
                   λ_prior  = (1.0, 1.0),
                   μ_prior  = (1.0, 1.0),
                   survival = true,
                   tρ       = Dict("" => 1.0))

where we now have a Gamma prior for the extinction $\mu$, and we can also specify if we want to condition on survival of the process with survival.

Full documentation

Tapestree.INSANE.insane_cbd — Function

insane_cbd(tree    ::sT_label;
           λ_prior ::NTuple{2,Float64}     = (1.0, 1.0),
           μ_prior ::NTuple{2,Float64}     = (1.0, 1.0),
           niter   ::Int64                 = 1_000,
           nthin   ::Int64                 = 10,
           nburn   ::Int64                 = 200,
           nflush  ::Int64                 = nthin,
           ofile   ::String                = string(homedir(), "/cbd"),
           ϵi      ::Float64               = 0.4,
           λi      ::Float64               = NaN,
           μi      ::Float64               = NaN,
           pupdp   ::NTuple{3,Float64}     = (0.2,0.2,0.2),
           prints  ::Int64                 = 5,
           survival::Bool                  = true,
           mxthf   ::Float64               = Inf,
           tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for constant birth-death.

Birth-death diffusion (BDD) processes

These models all assume that the per-lineage instantaneous speciation rates $\lambda(t)$ follow a Geometric Brownian motion (GBM), such that

\[d(\text{ln}(\lambda(t)) = \alpha dt + \sigma_{\lambda} d W(t),\]

where $\alpha$ is the drift, $\sigma_{\lambda}$ is the diffusion for speciation, and $W(t)$ is the Wiener process, and holding different assumptions on extinction.

The likelihood is approximated using the Euler method to solve Stochastic Differential Equations, and require to set a minimum time step for the discretization of the augmented diffusion paths. In INSANE δt (by default $10^{-3}$) sets the time step to perform the Euler approximation, and denotes the proportion with respect to the tree height (the duration of the tree), so the time step will be $δt = th \times 10^-3$, where $th$ is the tree height.

Pure birth diffusion ($\mu(t) = 0$)

In the simplest BDD model, the pure-birth diffusion, we assume there is no extinction, that is $\mu(t) = 0$.

Simulations

For all the BDD models, we have the possibility to simulate conditioned on total simulation time or the number of species.

To simulate conditioned on some number of species, say, $20$, with starting speciation rate of $\lambda_0 = 1.0$, drift of $\alpha = 0$ and diffusion of $\sigma_{\lambda} = 0.1$, we can use:

 sim_gbmpb(20, λ0 = 1.0, α = 0.0, σλ = 0.1)

Similarly, to simulate conditioned on time, say, $10$ time units, with the same parameters, we can use:

 sim_gbmpb(10.0, λ0 = 1.0, α = 0.0, σλ = 0.1)

Other options are available, such as δt which controls the time step of the simulation, which uses the Euler approximation. Similarly, init can be :crown or :stem to simulate starting with $1$ or $2$ lineages.

Note then that the simulations conditioned on time must input a Float64 as first (and mandatory) argument while those conditioned on number of species must rater input a Int64.

Full documentation

Tapestree.INSANE.sim_gbmpb — Function

sim_gbmpb(n       ::Int64;
          λ0      ::Float64 = 1.0,
          α       ::Float64 = 0.0,
          σλ      ::Float64 = 0.1,
          δt      ::Float64 = 1e-3,
          init    ::Symbol  = :stem,
          nstar   ::Int64   = n + 2,
          p       ::Float64 = 5.0,
          warnings::Bool    = true,
          maxt    ::Float64 = δt*1e7)

Simulate iTpb according to a pure-birth geometric Brownian motion.

sim_gbmpb(t   ::Float64;
          λ0  ::Float64 = 1.0,
          α   ::Float64 = 0.0,
          σλ  ::Float64 = 0.1,
          δt  ::Float64 = 1e-3,
          nlim::Int64   = 10_000,
          init::Symbol  = :crown)

Simulate iTpb according to a pure-birth geometric Brownian motion conditional in stopping at time t.

Inference

To perform inference under this model we can use:

r, tv = insane_gbmpb(tree,
                     nburn    = 1_000,
                     niter    = 50_000,
                     nthin    = 50, 
                     ofile    = "<directory>",
                     α_prior  = (0.0, 10.0),
                     σλ_prior = (0.05, 0.5),
                     tρ    = Dict("" => 1.0))

Here, the prior for the drift α_prior is a Normal distribution, with the first element representing the mean and the second the standard deviation, and the prior for the diffusion of speciation is an Inverse Gamma.

Full documentation

Tapestree.INSANE.insane_gbmpb — Function

insane_gbmpb(tree    ::sT_label;
             λ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             α_prior ::NTuple{2,Float64}     = (0.0, 1.0),
             σλ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             niter   ::Int64                 = 1_000,
             nthin   ::Int64                 = 10,
             nburn   ::Int64                 = 200,
             nflush  ::Int64                 = nthin,
             ofile   ::String                = string(homedir(), "/ipb"),
             αi      ::Float64               = 0.0,
             σλi     ::Float64               = 0.1,
             pupdp   ::NTuple{5,Float64}     = (0.01, 0.01, 0.01, 0.1, 0.2),
             δt      ::Float64               = 1e-3,
             prints  ::Int64                 = 5,
             stn     ::Float64               = 0.5,
             tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for pbd.

Birth-death diffusion process with constant extinction ($\mu(t) = \mu$)

Here we assume that there is extinction, but that is constant across lineages and time.

Simulations

As with the pure-birth diffusion, we can simulate conditioned on reaching some number of lineages or some time. Now, however, we need to specify the extinction rate as well, for instance to generate a tree with $20$ species and some parameter values:

 sim_gbmce(20, λ0 = 1.0, α = 0.0, σλ = 0.1, μ = 0.7)

Full documentation

Tapestree.INSANE.sim_gbmce — Function

sim_gbmce(n       ::Int64;
          λ0      ::Float64 = 1.0,
          α       ::Float64 = 0.0,
          σλ      ::Float64 = 0.1,
          μ       ::Float64 = 0.0,
          δt      ::Float64 = 1e-3,
          init    ::Symbol  = :stem,
          nstar   ::Int64   = 2*n,
          p       ::Float64 = 5.0,
          warnings::Bool    = true,
          maxt    ::Float64 = δt*1e7)

Simulate iTce according to a geometric Brownian motion for birth rates and constant extinction.

sim_gbmce(t   ::Float64;
          λ0  ::Float64 = 1.0,
          α   ::Float64 = 0.0,
          σλ  ::Float64 = 0.1,
          μ   ::Float64 = 0.2,
          δt  ::Float64 = 1e-3,
          nlim::Int64   = 10_000,
          init::Symbol  = :crown)

Simulate iTce according to a geometric Brownian motion for birth rates and constant extinction.

Inference

To perform inference we can use

r, tv = insane_gbmce(tree,
                     nburn    = 1_000,
                     niter    = 50_000,
                     nthin    = 50, 
                     ofile    = "<directory>",
                     α_prior  = (0.0, 10.0),
                     σλ_prior = (0.05, 0.5),
                     μ_prior  = (1.0, 1.0),
                     tρ       = Dict("" => 1.0))

where we now specify the Gamma prior μ_prior for the extinction.

Full documentation

Tapestree.INSANE.insane_gbmce — Function

insane_gbmce(tree    ::sT_label;
             λ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             α_prior ::NTuple{2,Float64}     = (0.0, 10.0),
             σλ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             μ_prior ::NTuple{2,Float64}     = (1.0, 1.0),
             niter   ::Int64                 = 1_000,
             nthin   ::Int64                 = 10,
             nburn   ::Int64                 = 200,
             nflush  ::Int64                 = nthin,
             ofile   ::String                = string(homedir(), "/ice"),
             λi      ::Float64               = NaN,
             αi      ::Float64               = 0.0,
             σλi     ::Float64               = 0.01,
             μi      ::Float64               = NaN,
             ϵi      ::Float64               = 0.2,
             pupdp   ::NTuple{5,Float64}     = (0.01, 0.01, 0.01, 0.1, 0.2),
             δt      ::Float64               = 1e-3,
             prints  ::Int64                 = 5,
             survival::Bool                  = true,
             mxthf   ::Float64               = 0.1,
             tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for gbm-ce.

Birth-death diffusion process with constant turnover ($\mu(t) = \epsilon \lambda(t)$)

One can also assume that turnover, $\epsilon$, that is, the ration of extinction over speciation (i.e., $\frac{\mu}{\lambda}$), is constant.

Simulations

As with the other BDD models, we can simulate both conditioned on the number of species or in time. For instance, here we condition on time with a turnover rate of $0.4$.

 sim_gbmct(10.0, λ0 = 1.0, α = 0.0, σλ = 0.1, ϵ = 0.4)

Full documentation

Tapestree.INSANE.sim_gbmct — Function

sim_gbmct(n       ::Int64;
          λ0      ::Float64 = 1.0,
          α       ::Float64 = 0.0,
          σλ      ::Float64 = 0.1,
          ϵ       ::Float64 = 0.0,
          δt      ::Float64 = 1e-3,
          init    ::Symbol  = :stem,
          nstar   ::Int64   = 2*n,
          p       ::Float64 = 5.0,
          warnings::Bool    = true,
          maxt    ::Float64 = δt*1e7)

Simulate iTct according to a geometric Brownian motion for birth rates and constant turnover.

sim_gbmct(t   ::Float64;
          λ0  ::Float64 = 1.0,
          α   ::Float64 = 0.0,
          σλ  ::Float64 = 0.1,
          ϵ   ::Float64 = 0.2,
          δt  ::Float64 = 1e-3,
          nlim::Int64   = 10_000,
          init::Symbol  = :crown)

Simulate iTct according to a geometric Brownian motion for birth rates and constant turnover.

Inference

Performing inference can then be done using:

r, tv = insane_gbmct(tree,
                     nburn    = 1_000,
                     niter    = 50_000,
                     nthin    = 50, 
                     ofile    = "<directory>",
                     α_prior  = (0.0, 10.0),
                     σλ_prior = (0.05, 0.5),
                     ϵ_prior  = (0.0, 100.0),
                     tρ       = Dict("" => 1.0))

where we now specify a Uniform prior ϵ_prior for the turnover. Also, INSANE resorts to simple MH updates for $\epsilon$, so an increased burnin is recommended.

Full documentation

Tapestree.INSANE.insane_gbmct — Function

insane_gbmct(tree    ::sT_label;
             λ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             α_prior ::NTuple{2,Float64}     = (0.0, 1.0),
             σλ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             ϵ_prior ::NTuple{2,Float64}     = (0.0, 10.0),
             niter   ::Int64                 = 1_000,
             nthin   ::Int64                 = 10,
             nburn   ::Int64                 = 200,
             nflush  ::Int64                 = nthin,
             ofile   ::String                = string(homedir(), "/ict"),
             tune_int::Int64                 = 100,
             αi      ::Float64               = 0.0,
             λi      ::Float64               = NaN,
             σλi     ::Float64               = 0.01,
             ϵi      ::Float64               = 0.2,
             ϵtni    ::Float64               = 0.1,
             obj_ar  ::Float64               = 0.234,
             pupdp   ::NTuple{5,Float64}     = (0.01, 0.01, 0.01, 0.1, 0.2),
             ntry    ::Int64                 = 2,
             nlim    ::Int64                 = 500,
             δt      ::Float64               = 1e-3,
             prints  ::Int64                 = 5,
             survival::Bool                  = true,
             mxthf   ::Float64               = Inf,
             tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for GBM birth-death.

Birth-death diffusion process

Finally, we arrive at the most complex model, where extinction also follows a GBM, that is,

\[d(\text{ln}(\mu(t)) = \sigma_{\mu} d W(t),\]

where $\sigma_{\mu}$ is the diffusion for extinction.

Note that this is unidentifiable, unless we specify strong priors on the extinction diffusion coefficient $\sigma_{\mu}$, and still do not do well in recovering simulated values, at least for small trees.

Simulations

Again, one can specify a time or a number of lineages to simulate under the BDD. For example, for $50$ species

 sim_gbmbd(50, λ0 = 1.0, μ0 = 1.0, α = 0.0, σλ = 0.1, σμ = 0.1)

Full documentation

Tapestree.INSANE.sim_gbmbd — Function

sim_gbmbd(n       ::Int64;
          λ0      ::Float64 = 1.0,
          μ0      ::Float64 = 0.1,
          α       ::Float64 = 0.0,
          σλ      ::Float64 = 0.1,
          σμ      ::Float64 = 0.1,
          init    ::Symbol  = :stem,
          δt      ::Float64 = 1e-3,
          nstar   ::Int64   = 2*n,
          p       ::Float64 = 5.0,
          warnings::Bool    = true)

Simulate iTbd according to geometric Brownian motions for birth and death rates.

sim_gbmbd(t   ::Float64;
          λ0  ::Float64 = 1.0,
          μ0  ::Float64 = 0.2,
          α   ::Float64 = 0.0,
          σλ  ::Float64 = 0.1,
          σμ  ::Float64 = 0.1,
          δt  ::Float64 = 1e-3,
          nlim::Int64   = 10_000,
          init::Symbol  = :crown)

Simulate iTbd according to geometric Brownian motions for birth and death rates.

Inference

One can perform inference using:

r, tv = insane_gbmbd(tree,
                     nburn    = 1_000,
                     niter    = 50_000,
                     nthin    = 50, 
                     ofile    = "<directory>",
                     λ0_prior = (0.05, 148.41),
                     μ0_prior = (0.05, 148.41),
                     α_prior  = (0.0, 10.0),
                     σλ_prior = (0.5, 0.05),
                     σμ_prior = (3.0, 0.5),
                     tρ       = Dict("" => 1.0))

where we have log-normal priors on the initial (root) values for $\lambda_0$ and $\mu_0$, and Inverse Gamma prior for sigma. Usually, an informative prior such as σμ_prior = (3.0, 0.5) or more is needed.

Full documentation

Tapestree.INSANE.insane_gbmbd — Function

insane_gbmbd(tree    ::sT_label;
             λ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             μ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             α_prior ::NTuple{2,Float64}     = (0.0, 10.0),
             σλ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             σμ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             niter   ::Int64                 = 1_000,
             nthin   ::Int64                 = 10,
             nburn   ::Int64                 = 200,
             nflush  ::Int64                 = nthin,
             ofile   ::String                = string(homedir(), "/ibd"),
             ϵi      ::Float64               = 0.2,
             λi      ::Float64               = NaN,
             μi      ::Float64               = NaN,
             αi      ::Float64               = 0.0,
             σλi     ::Float64               = 0.01,
             σμi     ::Float64               = 0.01,
             pupdp   ::NTuple{5,Float64}     = (0.01, 0.01, 0.0001, 0.1, 0.2),
             δt      ::Float64               = 1e-3,
             survival::Bool                  = true,
             mxthf   ::Float64               = 0.1,
             prints  ::Int64                 = 5,
             stnλ    ::Float64               = 0.5,
             stnμ    ::Float64               = 0.5,
             tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for bdd.

insane_gbmbd(tree    ::sT_label,
             tv      ::Vector{Vector{Float64}},
             ev      ::Vector{Vector{Float64}};
             λ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             μ0_prior::NTuple{2,Float64}     = (0.05, 148.41),
             α_prior ::NTuple{2,Float64}     = (0.0, 0.5),
             σλ_prior::NTuple{2,Float64}     = (3.0, 0.5),
             σμ_prior::NTuple{2,Float64}     = (5.0, 0.5),
             niter   ::Int64                 = 1_000,
             nthin   ::Int64                 = 10,
             nburn   ::Int64                 = 200,
             nflush  ::Int64                 = nthin,
             ofile   ::String                = string(homedir(), "/ibd_efx"),
             ϵi      ::Float64               = 0.2,
             λi      ::Float64               = NaN,
             αi      ::Float64               = 0.0,
             σλi     ::Float64               = 0.01,
             σμi     ::Float64               = 0.01,
             pupdp   ::NTuple{4,Float64}     = (0.01, 0.01, 0.1, 0.2),
             δt      ::Float64               = 1e-3,
             prints  ::Int64                 = 5,
             survival::Bool                  = true,
             mxthf   ::Float64               = Inf,
             tρ      ::Dict{String, Float64} = Dict("" => 1.0))

Run insane for gbm-bd with fixed extinction.

Birth-death diffusion process with informed extinction

One can also define a branch-specific fixed extinction function and make inference on the GBM speciation. For example, in Quintero et al. (2024) Science, we estimated extinction rates based on fossil occurrences and then used this as input.

Inference

To make inference under this model, we need two vectors where each element links to a given branch in the reconstructed tree, in the same order as the phylogenetic tree. Each element of these vectors are themselves a vector of type Float64, and correspond to the time at which extinction is recovered in time backward order (e.g., [1.1, 1.0, ..., 0.1, 0.08, 0.0]) and to the extinction at those times. INSANE uses a linear approximation function between the input sampled points.

To make inference under this model, we now input these vectors. Lets call them time_vector (of type Vector{Vector{Float64}}) and extinction_vector (also of type Vector{Vector{Float64}}), then:

r, tv = insane_gbmbd(tree,
                     time_vector,
                     extinction_vector,
                     nburn    = 1_000,
                     niter    = 50_000,
                     nthin    = 50, 
                     ofile    = "<directory>",
                     tρ       = Dict("" => 1.0))

The only thing one has to make sure is that the order of time_vector and extinction_vector correspond to the order of the branches in tree, which can be a bit involved. However, this can be achieved by understanding that the trees are recursive and are ordered either by left or right branch, and using some of the tree utility functions.

For example, one can use the subclade function, to extract a subclade given a set of tip labels. One can also use the make_idf function, which splits the tree into a vector of branches in the same order that INSANE will process the tree. This way one can see the ordering of branches, and then order the time_vector and the extinction_vector in the same order.

If only one extinction function is used for all the tree, then one has to only create a vector of the time and extinction vector of equal size as the number of branches.

To estimate the number of branches in your empirical tree (saved in <...tree directory...>), we can do:

# read tree
tre = read_newick("<...tree directory...>")

## make extinction function vectors for each branch
# set a dummy smapling fraction (you can set the real one afterwards! - this is just to estimate the number of branches)
tρ  = Dict(li => 1.0 for li in tiplabels(tre))
idf = make_idf(tre, tρ, Inf)

# number of branches
nb = length(idf)

If we have a global curve with extinction [0.06, 0.02, 0.05] at times [0.5, 0.3, 0.1], then we would have to create the following input objects for time_vector and extinction_vector:

extinction_vector = fill([0.06, 0.02, 0.05], nb)
time_vector = fill([0.5, 0.3, 0.1], nb)

Diffused Brownian motion process (DBM)

This is a model of trait evolution where an univariate trait $x(t)$ evolves under a diffused Brownian motion with an underlying evolutionary rate $\sigma^2(t)$ that is also itself evolving separately according to Geometric Brownian motion.

we arrive at the most complex model, where extinction also follows a GBM, that is,

\[dx(t) = \alpha_x dt + \sigma(t) d W(t) \\ d(\text{ln}(\sigma^2(t)) = \alpha_{\sigma} dt + \gamma d W(t)\]

where $\alpha_x$ is the trait drift (general trait tendency to increase or decrease), $\alpha_{\sigma}$ is the drift in evolutionary rates, and $\gamma$ represents the heterogeneity in evolutionary rates.

Simulations

In the DBM, we are not modeling the realization of the tree, but rather a process that evolved along the tree. Thus, to simulate, we need a tree to be used as template so that we can simulate on top. So if we have a tree object of type sT_label or sTf_label, we can simulate a trait with a starting trait value of x0 ($x(t = 0)$), with drift αx ($\alpha_x$), undergoing a starting rate of σ20 ($\sigma^2(t = 0)$) with drift ασ ($\alpha_{\sigma}$), with a maximal discretization time step of δt:

sim_dbm(tree, x0 = 0.0, αx = 0.0, σ20 = 0.1, ασ = 0.0, γ = 0.1, δt = 1e-3)

which returns a tree of type sTxs, holding the simulation.

It might also be useful to return a Dictionary with the final trait values at each sampled species. For this, use the same function, in the same argument order but not using named arguments. For instance, the same simulation above, but returning both a sTxs tree named tr and a dictionary of species values named xd:

tr, xd = sim_dbm(tree, 0.0, 0.0, 0.1, 0.0, 0.1, 1e-3)

We can plot the resulting tree using Tapestree's plot recipes (Insane plots). For example to plot the trait evolution colored by the logarithmic rates:

plot(xv, tr, zf = lσ2)

Full documentation

Tapestree.INSANE.sim_dbm — Function

sim_dbm(tree::iTree; 
        x0  ::Float64 = 0.0,
        αx  ::Float64 = 0.0,
        σ20 ::Float64 = 0.1,
        ασ  ::Float64 = 0.0,
        γ   ::Float64 = 0.1,
        δt  ::Float64 = 1e-3)

Simulate a diffused Brownian motion given starting values.

sim_dbm(tree::Tlabel, 
        x0  ::Float64,
        αx  ::Float64,
        σ20 ::Float64,
        ασ  ::Float64,
        γ   ::Float64,
        δt  ::Float64)

Simulate a diffused Brownian motion given starting values.

Inference

For a given sT_label or sTf_label type tree object and a Dictionary xav with a String key pointing to a Float64 number (i.e., Dict{String, Float64}), where matching tip labels point to the trait value. If tip labels are not included in the dictionary, the trait value is assumed missing.

r, td = insane_dbm(tree, xav,
                   γ_prior  = (0.05, 0.05),
                   αx_prior = (0.0, 10.0),
                   ασ_prior = (0.0, 10.0),
                   nburn    = 10_000,
                   niter    = 100_000,
                   nthin    = 1_000,
                   nflush   = 1_000,
                   ofile    = "<...directory...>",
                   δt       = 1e-3)

Finally, error or uncertainty around trait values can be included (assuming Normal variance) by setting another Dictionary, called say xsv, (also Dict{String, Float64}), where tip values point to the variance around trait values. Again, if tip labels are not included in this dictionary, it is assumed that there is no error around tip values. Then you specify this dictionary on the argument xs:

r, td = insane_dbm(tree, xav, 
                   xs       = xsv,
                   γ_prior  = (0.05, 0.05),
                   αx_prior = (0.0, 10.0),
                   ασ_prior = (0.0, 10.0),
                   nburn    = 10_000,
                   niter    = 100_000,
                   nthin    = 1_000,
                   nflush   = 1_000,
                   ofile    = "<...directory...>",
                   δt       = 1e-3)

Full documentation

Tapestree.INSANE.insane_dbm — Function

insane_dbm(tree     ::Tlabel,
           xa       ::Dict{String, Float64};
           xs       ::Dict{String, Float64} = Dict{String,Float64}(),
           αx_prior ::NTuple{2,Float64} = (0.0, 10.0),
           ασ_prior ::NTuple{2,Float64} = (0.0, 10.0),
           γ_prior  ::NTuple{2,Float64} = (0.05, 0.05),
           niter    ::Int64             = 1_000,
           nthin    ::Int64             = 10,
           nburn    ::Int64             = 200,
           nflush   ::Int64             = nthin,
           ofile    ::String            = string(homedir(), "/dbm"),
           αi       ::Float64           = 0.0,
           γi       ::Float64           = 1e-3,
           pupdp    ::NTuple{4,Float64} = (0.1, 0.1, 0.05, 0.9),
           δt       ::Float64           = 1e-3,
           stn      ::Float64           = 0.1,
           mxthf    ::Float64           = Inf,
           prints   ::Int64             = 5)

Run diffused Brownian motion trait evolution model.

Insane tree data access and processing functions

Simple tree information

Simple information about the tree, such as the number tips, the number of extinct tips, the tree height (duration of the tree) and the tree length (sum of all branch lengths) can be performed using, respectively:

ntips(tree)
ntipsextinct(tree)
treeheight(tree)
treelength(tree)

Julia makes it simple to look at statistics across a vector of trees. For example, using the package Statistics, we can estimate the average number of extinct species on tree vector tv by simply:

using Statistics 

mean(ntipsextinct, tv)

See the documentation of mean for more details, but basically, mean, and many other functions in Julia allow to perform a undefined function on each element before calculating the mean. In this case, we are estimating the number of extinct tips in each tree in tv, and then averaging over them.

For labelled trees, one can extract the tip labels using tiplabels. Moreover we can create subclades based on a vector of tips, where the subclade will be the minimum tree that has all the tips. For example, for a vector tip_vector holding Strings that correspond to the tip labels in the tree of type sT_label, we can use

subclade(tree, tip_vector)

However, most time we want to extract subclades of other types of trees. Since these do not hold label information but should be ordered in the same order as the sT_label tree, one has to use both. Note, if you change the order of either the sT_label tree or the single of vector of trees of other types, this will not work. Thus, if we want a the subclades that have the tips in tip_vector, we can use

subclade(tv, tree, tip_vector, true)

where the last argument states if returning the stem or crown tree.

One can also estimate the Lineage Through Time (LTT), or, for a data augmented tree (or a vector of trees), the Diversity Through Time (DTT) using

ltt(tree)

To be clear, the LTT is usually used to describe the accumulation of reconstructed lineages (those that have been sampled) while DTT is used to describe estimated diversity (sampled and unsampled lineages). Thus, the result is either LTT or DTT simply depending on the tree you use as input.

We can also estimate the ltt for a tree vector tv using

ltt(tv)

BDD information

Estimating posterior average rates along the tree

Of particular interest is the estimation of posterior average rates along the reconstructed tree. Note that since the data augmented (unsampled) lineages change between different iterations of the algorithm, we obtain lineage-specific instantaneous rate distributions only for the reconstructed (observed) tree (the tree we used as input). Consequently, we first need to remove the data augmented lineages from all the trees in the posterior tree vector:

tv0 = remove_unsampled(tv)

We can then estimate the average tree using

tm = imean(tv0)

We can also estimate any quantile tree, for instance, for the $0.25$ quantile tree:

t025 = iquantile(tv0, 0.25)

Clearly, these resulting trees can then be further scrutinized as with any other tree in INSANE.

Other data access and averaging functions

If one wants to obtain the range (i.e., extrema) of the output of function f on tree, for example, the maximum and minimum speciation rates:

irange(tree, b)

If one wants to sample, recursively, some function at regular intervals along a tree, one can use sample. For example if we want to sample speciation rates every $0.1$ time units, we can use

sample(tv, b, 0.1)

Note that here we are sampling along each branch of the tree in recursive order, not sampling across lineages through time. If we would like to extract an array across lineages in a given tree of the output of function f, we would use time_rate. For example, if we want the cross-lineage extinction rates of a tree of type iTbd sampled every $0.5$ time units, we would use

time_rate(tv, d, 0.5)

which returns a vector of vectors, where each element is a time holding the rates (in this case extinction rates) of all contemporary lineages at that time.

Finally, a convenience wrapper to extract information recursively from a tree is trextract. For example, if we want all branch lengths for a tree, we can use

trextract(tree, e)

Below are some functions to obtain data from trees.

Full documentation

Tapestree.INSANE.tiplabels — Function

tiplabels(tree::T) where {T <: Tlabel}

Returns tip labels for sT_label and sTf_label.

Tapestree.INSANE.ntips — Function

ntips(tree::T) where {T <: iTree}

Return the number of tip nodes for tree.

Tapestree.INSANE.ntipsalive — Function

ntipsalive(tree::T) where {T <: iTree}

Return the number of alive nodes for tree.

Tapestree.INSANE.ntipsextinct — Function

ntipsextinct(tree::T) where {T <: iTree}

Return the number of extinct nodes for tree.

ntipsextinct(Ξ::Vector{T}) where {T <: iTree}

Return the number of extinct nodes in Ξ.

Tapestree.INSANE.treeheight — Function

treeheight(tree::T) where {T <: iTree}

Return the tree height of tree.

treeheight(tree::T) where {T <: iTf}

Return the tree height of tree.

treeheight(tree::T, nd::Int64) where {T <: iTree}

Return the tree height of tree.

treeheight(tree::T, nd::Int64) where {T <: iTf}

Return the tree height of tree.

Tapestree.INSANE.treelength — Function

treelength(tree::T) where {T <: iTree}

Return the branch length sum of tree.

treelength(tree::T, ets::Vector{Float64})  where {T <: iTf}

Return the branch length sum of tree at different epochs, initialized at l.

treelength(Ξ::Vector{T}) where {T <: iTree}

Return the branch length sum of Ξ.

treelength(Ξ  ::Vector{T},
           ets::Vector{Float64},
           bst::Vector{Float64},
           eix::Vector{Int64})  where {T <: iTf}

Return the branch length sum of tree at different epochs, initialized at l.

Tapestree.INSANE.ltt — Function

ltt(tree::T) where {T <: iTree}

Returns number of species through time.

ltt(tree::Vector{T}) where {T <: iTree}

Returns number of species through time for a tree vector.

Tapestree.INSANE.iscrowntree — Function

iscrowntree(tree::T) where {T <: iTree}

Return if the tree is a crown tree.

Tapestree.INSANE.irange — Function

irange(tree::T, f::Function) where {T <: iTree}

Return the extrema of the output of function f on tree.

Tapestree.INSANE.time_rate — Function

time_rate(tree::T, f::Function, δt::Float64) where {T <: iT}

Extract values from f function at times sampled every δt across the tree.

Tapestree.INSANE.trextract — Function

trextract(tree::iTree, f::Function)

Perform function f in each recursive tree in tree.

Tapestree.INSANE.subclade — Function

subclade(tree::iTree, ix::Int64)

Return the minimum stem subclade according to recursive position ix.

subclade(trees::Vector{T}, 
              ltree::sT_label, 
              tips ::Vector{String},
              stem ::Bool) where {T <: iTree}

Return the minimum subclade that includes tip labels in tips.

subclade(tree::sT_label, tips::Vector{String})

Return the minimum subclade that includes tip labels in tips.

subclade(tree::iTree, 
         ltree::sT_label, 
         tips ::Vector{String}, 
         stem ::Bool)

Return the minimum subclade that includes tip labels in tips.

Tapestree.INSANE.lλ — Function

lλ(tree::T) where {T <: iT}

Return pendant edge.

Tapestree.INSANE.lμ — Function

lμ(tree::iTbdU)

Return pendant edge.

Tapestree.INSANE.e — Function

e(id::iB)

Return initial absolute time.

e(tree::T) where {T <: iTree}

Return edge length.

Insane tree manipulation functions

Two important manipulation functions are, first to be able to remove extinct lineages, which can be performed on a tree or a tree vector using

remove_extinct(tree)

Similarly, as shown above, one can remove the unsampled lineages (all the data augmented lineages) on a single or vector of trees using

remove_unsampled(tree)

Note that these two are different. First, when performing simulations, the tree is not fixed, which means that if you run remove_unsampled, you will remove the tree. You would have to fix the tree before, which can be done using fixtree!(tree). Also, if sampling fraction is not $1$, remove_unsampled will also remove lineages alive that were not sampled, while remove_extinct will only remove those lineages extinct.

Full documentation

Tapestree.INSANE.reorder! — Function

reorder(tree::T) where {T <: iTree}

Reorder order of daughter branches according to number of tips, with daughter 1 always having more than daughter 2.

reorder!(tree::T, treeda::D) where {T <: iTree, D <: iTree}

Reorder order of daughter branches for both trees, following tree first, according to number of tips, with daughter1 always having more than daughter 2.

Tapestree.INSANE.rm_stem! — Function

rm_stem(tree::T) where {T <: iTree}

Removes stem branch.

Tapestree.INSANE.fixtree! — Function

fixtree!(tree::T) where {T <: iTree}

Fix all tree.

fixtree!(tree::T) where {T <: iTf}

Fix all tree.

Tapestree.INSANE.remove_extinct — Function

remove_extinct(tree::T) where {T <: iTree}

Remove extinct tips from iTce.

remove_extinct(treev::Vector{T}) where {T <: iTree}

Remove extinct taxa for a vector of trees.

Tapestree.INSANE.remove_unsampled — Function

remove_unsampled(tree::T) where {T <: iTree}

Remove unsampled tips from iTpb.

remove_unsampled(treev::Vector{T}) where {T <: iTree}

Remove unsampled taxa for a vector of trees.

Insane plots

Tapestree holds recipes to plot phylogenetic trees, results and perform aggregating functions. The output will rely in the number and type of arguments you provide.

First, however, we need to load the Plots package

using Plots

Insane plots across all tree types

Simple tree plot

The most basic plot function simply plots the tree:

plot(tree)

One can reorder the tree according to balance (have one daughter always have the largest number of tips) by using reorder!(tree), which orders in place the tree and helps in visualization.

If one has a tree vector, we could, for example, sample 4 of them at random and simply plot them together using

ti = rand(tv,4)

p0 = plot(ti[1])
p1 = plot(ti[2])
p2 = plot(ti[3])
p3 = plot(ti[4])

plot(p0, p1, p2, p3)

If the tree is of type sT_label, labels will be shown automatically, but you can toggle this off with showlabels = false.

One can also plot the tree radially (as a fan) using

plot(tree, type = :radial)

LTT and DTT plots

We can plot the LTT or DTT by using the ltt result (of type Ltt, check (check Simple tree information)) as input:

plot(ltt(tree), linewidth = 2.0)

Moreover, if we input a vector of Ltt we will plot each LTT individually, or, better, if we add a decimal number argument, it will use it as sampling frequency through time and return the mean and desired quantiles of lineage or diversity through time using the arguments q0 and q1 (by default q0 = [0.025, 0.975] and q1 = [0.25, 0.75]):

lttv = ltt(tv)
plot(lttv, 0.1)

Insane BDD & DBM plots

These plotting functions are specific to BDD or DBM type trees (i.e., of iT or sTxs supertype).

Plot rates on the phylogram

To "paint" the tree with the instantaneous lineage-specific rates of speciation $\lambda(t)$, we can use:

plot(tree, birth)

Here, birth, which stands for "birth rates", is a convenience wrapper around exp.(lλ(tree)): it extracts the log-speciation vector from a give iT tree using lλ, and then returns the exponential.

This plotting function also allows to plot the death rates (only where extinction is also a diffusion, i.e., iTbd) using

plot(tree, death)

where death, which stands for "death rates" is a wrapper around exp.(lμ(tree)).

In general, this plotting recipe receives a tree and a function that is applied recursively to paint the tree. Thus, we can use any custom made function that extracts information from the tree. Some predefined ones are:

logbirth: log-(speciation) birth rates

logdeath: log-(extinction) death rates

turnover: turnover (extinction/speciation) rates

diversification: net diversification (speciation - extinction) rates

trait: trait values

evorate: evolutionary rates $\sigma^2(t)$

logevorate: log-evolutionary rates $\sigma^2(t)$

We can also plot these trees radially using the type = :radial.

Plot the underlying rates along a tree

To plot how rates evolve across time, that is, to plot the rates in the y axis, we simply change the argument order:

plot(birth, tree)

We can paint these rates by another variable, say, extinction rates, using

plot(birth, tree, zf = death)

Similarly, we can plot the average for a tree (or other aggregating function as median, geometric mean, etc.) and custom quantiles (as in LTT and DTT plots) for a given tree by adding a decimal number argument representing the sampling frequency through time.

plot(birth, 0.1, tree)

To change the aggregating function, we modify the function t_af (by default t_af = mean), to the desired one.

Plot the rates across tree vectors

Often we would like to plot the average rates across a series of data augmented trees. This can be done by adding a decimal number argument (and, using a tree vector as input). For instance, to estimate average speciation rates (using wrapping function birth) through time across tree vector tv, every $0.1$ time units, we use:

plot(birth, 0.1, tv)

We can choose the function to aggregate rates across lineages for each single tree using af (by default af = mean, and then to aggregate these tree averages using the vaf function (by default vaf = x -> quantile(x, 0.5), that is, the median).